requirements.txt
ansicon==1.89.0
bidict==0.22.0
blessed==1.19.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Login==0.6.2
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tzdata==2022.7
tzlocal==4.2
wcwidth==0.2.5
Werkzeug==2.2.2
